ATLASNAK-26 Make sure users don't end up being members and managers simultaneously.
There's some code in LiteAbstractSakaiGroupPostServlet whose job it is to remove a user from the member list when they get added to the managers list (and vice versa).  At the heart of this is a method called "getPeerGroupOf" which gives you "foo-members" if you give it "foo", and "foo" if you give it "foo-members".  The logic was basically:

  if (group.hasPropery("managers-group")) {
    // Return the managers group of the current group
  } else if (group.hasProperty("managed-group")) {
    // Return the group that this group manages
  }

At some point along the line, manager groups ended up being their own managers (whoa.) so getPeerGroupOf("some-group-managers") would always return "some-group-managers".  Not really what we wanted.

So I've modified getPeerOf to flip the order of the if branches--giving the "managed-group" property precedence.